Refactor: move to common directory
