--- src/components/pages/main/StyledLink.jsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { Link } from 'react-router-dom';
-
-const StyledLink = styled(Link)`
-  text-decoration: none;
-`;
-
-export default StyledLink;
